Business Rules are written

+++ src/main/java/kodlama/io/rentacar/business/concretes/BrandManager.java
@@ -6,6 +6,7 @@ import kodlama.io.rentacar.business.requests.UpdateBrandRequest;
 import kodlama.io.rentacar.business.responses.GetAllBrandsResponse;
 import kodlama.io.rentacar.business.responses.GetByIdBrandResponse;
 import kodlama.io.rentacar.business.rules.BrandBusinessRules;
+import kodlama.io.rentacar.core.utilities.exceptions.BusinessException;
 import kodlama.io.rentacar.core.utilities.mappers.ModelMapperService;
 import kodlama.io.rentacar.dataAccess.abstracts.BrandRepository;
 import kodlama.io.rentacar.entities.concretes.Brand;
@@ -58,7 +59,13 @@ public class BrandManager implements BrandService {
 
     @Override
     public void delete(int id) {
-        this.brandRepository.deleteById(id);
+        if(brandRepository.existsById(id)){
+            this.brandRepository.deleteById(id);
+        }else {
+            throw new BusinessException("aynı id silinemez");
+        }
+
+
 
     }
 }
